+++ TeamProject/Controllers/AuthControllers/ResetPasswordViewController.swift
@@ -17,7 +17,6 @@ class ResetPasswordViewController: UIViewController {
         super.viewDidLoad()
         configureUI()
         setupButtonActions()
-        
     }
